Rename serialization data providers
